% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mccall93}
\alias{data.mccall93}
\title{Data from a meta-analysis infants' habituation to a give stimulus and their later cognitive
ability (IQ)}
\format{The data frame contains two columns: ri (raw correlation coefficients) and ni (sample sizes).}
\source{
McCall, R. B., & Carriger, M. S. (1993). A meta-analysis of infant habituation and recognition
memory performance as predictors of later IQ. Child Development, 64, 57-79.
http://dx.doi.org/10.2307/1131437
}
\description{
Results from 12 studies on the association between infants' habituation to a give stimulus and
their later cognitive ability (IQ).
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}
\keyword{data}
