% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GIT_escompute_3.R
\name{escompute}
\alias{escompute}
\title{escompute}
\usage{
escompute(mi, ri, ni, sdi, m1i, m2i, n1i, n2i, sd1i, sd2i, tobs, alpha, side,
  measure)
}
\arguments{
\item{mi}{A vector of group means for one-sample mean}

\item{ri}{A vector of raw correlation coefficients}

\item{ni}{A vector of sample sizes for one-sample mean}

\item{sdi}{A vector of standard deviations for one-sample mean}

\item{m1i}{A vector of means in group 1 for two-sample mean}

\item{m2i}{A vector of means in group 2 for two-sample mean}

\item{n1i}{A vector of sample sizes in group 1 for two-sample mean}

\item{n2i}{A vector of sample sizes in group 2 for two-sample mean}

\item{sd1i}{A vector of standard deviations in group 1 for two-sample mean}

\item{sd2i}{A vector of standard deviations in group 2 for two-sample mean}

\item{tobs}{A vector of t-values}

\item{alpha}{A integer specifying the alpha level as used in primary studies}

\item{side}{A character indicating the direction of the tested hypothesis in the primary studies (either "right" or "left")}

\item{measure}{A character indicating what kind of effect size should be computed (Hedges' g or Fisher's r-to-z transformed correlation coefficients) and which arguments are used as input ("M", "MT", "MD", "MDT", or "COR"). See Details below.}
}
\value{
Function returns a data frame with standardized effect sizes (yi), variances of these standardized effect sizes (vi), z-values (zval), p-values as computed in primary studies (pval), and critical z-values (zcv).
}
\description{
Function that computes Hedges' g for an one-sample mean and a two-sample mean, Fisher's r-to-z transformed correlation coefficient for a raw correlation coefficient and computes a p-value as in primary studies.
}
\details{
The \code{measure} argument has to be used to specify the desired effect size and what input parameters are used. There are five options:
\itemize{
\item{\code{"M"}}{ for one-sample mean with \code{mi}, \code{ni}, \code{sdi}, \code{alpha}, and \code{side} as input parameters}
\item{\code{"MT"}}{ for one-sample mean with \code{tobs}, \code{ni}, \code{alpha}, and \code{side} as input parameters}
\item{\code{"M"}}{ for two-sample mean with \code{m1i}, \code{m2i}, \code{n1i}, \code{n2i}, \code{sd1i}, \code{sd2i}, \code{alpha}, and \code{side} as input parameters}
\item{\code{"MT"}}{ for two-sample mean with \code{tobs}, \code{n1i}, \code{n2i}, \code{alpha}, and \code{side} as input parameters}
\item{\code{"COR"}}{ for raw correlation coefficients with \code{ri}, \code{ni}, \code{alpha}, and \code{side} as input parameters}
}
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}

