% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.ma.R
\name{fe.ma}
\alias{fe.ma}
\title{fe.ma}
\usage{
fe.ma(yi, vi)
}
\arguments{
\item{yi}{A vector of standardized effect sizes}

\item{vi}{A vector of sampling variances belonging to the standardized effect sizes (\code{yi})}
}
\value{
Function returns a data frame with the following results of fixed-effect meta-analysis:
estimate (\code{est.fe}), standard error (\code{se.fe}), lower bound 95\% confidence interval (\code{ci.lb.fe}),
upper bound 95\% confidence interval (\code{ci.ub.fe}), z-value of test of no effect (\code{zval.fe}),
two-tailed p-value of test of no effect (\code{pval.fe}), one-tailed p-value of test of no effect (\code{pval.fe.one}),
test statistic of the Q-test (\code{Qstat}), and p-value of the Q-test (\code{Qpval}).
}
\description{
Function that conducts fixed-effect meta-analysis
}
\details{
The \code{fe.ma} function can be used for conducting fixed-effect meta-analysis
on a set of studies
}

