% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_ma.R
\name{fe_ma}
\alias{fe_ma}
\title{fe_ma}
\usage{
fe_ma(yi, vi)
}
\arguments{
\item{yi}{A vector of standardized effect sizes}

\item{vi}{A vector of sampling variances belonging to the standardized effect sizes (\code{yi})}
}
\value{
\item{est.fe}{effect size estimate of fixed-effect meta-analysis}
\item{se.fe}{standard error of estimate of fixed-effect meta-analysis}
\item{ci.lb.fe}{lower bound 95\% confidence interval}
\item{ci.ub.fe}{upper bound 95\% confidence interval}
\item{zval.fe}{z-value of test of no effect}
\item{pval.fe}{two-tailed p-value of test of no effect}
\item{pval.fe.one}{one-tailed p-value of test of no effect}
\item{Qstat}{test statistic of the Q-test}
\item{Qpval}{p-value of the Q-test}
}
\description{
Function that conducts fixed-effect meta-analysis
}
\details{
The \code{fe.ma} function can be used for conducting fixed-effect meta-analysis
on a set of studies
}
