% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis.trans.R
\name{fis.trans}
\alias{fis.trans}
\title{fis.trans}
\usage{
fis.trans(r, fis)
}
\arguments{
\item{r}{An integer being a raw correlation coefficient}

\item{fis}{An integer being a Fisher-z transformed correlation coefficient}
}
\description{
Function for transforming raw correlation coefficients to Fisher-z transformed
correlation coefficients and vice versa.
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}
