% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hybridoutput.R
\name{print.hybridoutput}
\alias{print.hybridoutput}
\title{print.hybridoutput}
\usage{
\method{print}{hybridoutput}(x, ...)
}
\arguments{
\item{x}{an object of class \code{hybridoutput}}

\item{...}{Additional arguments that can be passed to the function}
}
\value{
The \code{print.hybridoutput} function does not return an object.
}
\description{
Print method for objecs of class \code{hybridoutput}
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}
