% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.puni_staroutput.R
\name{print.puni_staroutput}
\alias{print.puni_staroutput}
\title{print.puni_staroutput}
\usage{
print.puni_staroutput(x, ...)
}
\arguments{
\item{x}{an object of class \code{puni_staroutput}}

\item{...}{Additional arguments that can be passed to the function}
}
\value{
The \code{print.puni_staroutput} function does not return an object.
}
\description{
Print method for objecs of class \code{puni_staroutput}
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}
