% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GIT_puniform_compleet_9.R
\name{puniform}
\alias{puniform}
\title{p-uniform}
\usage{
puniform(mi, ri, ni, sdi, m1i, m2i, n1i, n2i, sd1i, sd2i, tobs, alpha = 0.05,
  side, method, plot = FALSE)
}
\arguments{
\item{mi}{A vector of group means for one-sample means}

\item{ri}{A vector of raw correlations}

\item{ni}{A vector of sample sizes for one-sample means and correlations}

\item{sdi}{A vector of standard deviations for one-sample means}

\item{m1i}{A vector of means in group 1 for two-sample means}

\item{m2i}{A vector of means in group 2 for two-sample means}

\item{n1i}{A vector of sample sizes in group 1 for two-sample means}

\item{n2i}{A vector of sample sizes in group 2 for two-sample means}

\item{sd1i}{A vector of standard deviations in group 1 for two-sample means}

\item{sd2i}{A vector of standard deviations in group 2 for two-sample means}

\item{tobs}{A vector of t-values}

\item{alpha}{A integer specifying the alpha level as used in primary studies (default is 0.05)}

\item{side}{A character indicating the direction of the tested hypothesis in the primary studies (either \code{"right"} or \code{"left"})}

\item{method}{A character indicating the method to be used (\code{"P"} (default), \code{"LNP"}, \code{"LN1MINP"}, \code{"KS"}, or \code{"AD"})}

\item{plot}{A logical indicating whether a plot showing the relation between observed and expected p-values has to be rendered (default is \code{TRUE})}
}
\value{
\item{est}{p-uniform's effect size estimate}
\item{ci.lb}{lower bound of p-uniform's confidence interval}
\item{ci.ub}{upper bound of p-uniform's confidence interval}
\item{ksig}{number of significant studies}
\item{L.0}{test statistic of p-uniform's test of null-hypothesis of no effect (for method \code{"P"} a z-value)}
\item{pval.0}{one-tailed p-value of p-uniform's test of null-hypothesis of no effect}
\item{L.pb}{test statistic of p-uniform's publication bias test}
\item{pval.pb}{one-tailed p-value of p-uniform's publication bias test}
\item{est.fe}{effect size estimate based on traditional fixed-effect meta-analysis}
\item{se.fe}{standard error of effect size estimate based on traditional fixed-effect meta-analysis}
\item{zval.fe}{test statistic of the null-hypothesis of no effect based on traditional fixed-effect meta-analysis}
\item{pval.fe}{one-tailed p-value of the null-hypothesis of no effect based on traditional fixed-effect meta-analysis}
\item{ci.lb.fe}{lower bound of confidence interval based on traditional fixed-effect meta-analysis}
\item{ci.ub.fe}{ci.ub.fe upper bound of confidence interval based on traditional fixed-effect meta-analysis}
\item{Qstat}{test statistic of the Q-test for testing the null-hypothesis of homogeneity}
\item{Qpval}{one-tailed p-value of the Q-test}
}
\description{
Function to apply p-uniform method for one-sample means, two-sample means, and raw correlation coefficients. \cr
\cr
Please note that the method is still in development and that this is a beta version. If you suspect a bug, please send me an email (\email{R.C.M.vanAert@tilburguniversity.edu}).
}
\details{
Three different effect sizes can be used as input for the \code{puniform} function: one-sample means, two-sample means, and raw correlation coefficients. Analyzing one-sample
means and two-sample means can be done by either providing the function group means (\code{mi} or \code{m1i} and \code{m2i}), standard deviations (\code{sdi} or \code{sd1i} and \code{sd2i}),
and sample sizes (\code{ni} or \code{n1i} and \code{n2i}) or t-values (\code{tobs}) and sample sizes (\code{ni} or \code{n1i} and \code{n2i}). Both options should be accompanied with input for
the arguments \code{side} and \code{method}. See the Example section for examples.

Raw correlation coefficients can be analyzed by supplying \code{ri} and \code{ni} to the \code{puniform} function next to input for the arguments \code{side} and \code{method}.

Note that only one effect size measure can be specified at a time. A combination of effect size measures usually causes true heterogeneity among effect sizes and including different effect size measures
is therefore not recommended.
}
\note{
The p-uniform method is described in: \cr
\cr
van Assen, M. A. L. M., van Aert, R. C. M., & Wicherts, J. M. (2014).
Meta-analysis using effect size distributions of only statistically significant studies. Psychological Methods,
Advance online publication. doi: http://dx.doi.org/10.1037/met0000025
}
\examples{
### Load data from meta-analysis by McCall and Carriger (1993)
data(data.mccall93)

### Apply p-uniform method to get the same results as in van Assen et al. (2014)
puniform(ri = data.mccall93$ri, ni = data.mccall93$ni, alpha = .025, side = "right", method = "LNP", plot = TRUE)

### Note that the results of the publication bias test of p-uniform are not exactly equal to the results as stated in van Assen et al. (2014).
### This is caused by a minor mistake in the analyses in van Assen et al. (2014).

### Generate some example data
set.seed(123)
ni <- 100
sdi <- 1
mi <- rnorm(8, mean = 0.2, sd = sdi/sqrt(ni))
tobs <- mi/(sdi/sqrt(ni))

### Apply p-uniform method based on sample means
puniform(mi = mi, ni = ni, sdi = sdi, alpha = 0.05, side = "right", method = "P", plot = FALSE)

### Apply p-uniform method based on t-values
puniform(ni = ni, tobs = tobs, alpha = 0.05, side = "right", method = "P", plot = FALSE)
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}

