% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GIT_escompute_1.R
\name{escompute}
\alias{escompute}
\title{escompute}
\usage{
escompute(mi, ri, ni, sdi, m1i, m2i, n1i, n2i, sd1i, sd2i, side, measure)
}
\arguments{
\item{mi}{A vector of group means for one-sample mean}

\item{ni}{A vector of sample sizes for one-sample mean}

\item{sdi}{A vector of standard deviations for one-sample mean}

\item{m1i}{A vector of means in group 1 for two-sample mean}

\item{m2i}{A vector of means in group 2 for two-sample mean}

\item{n1i}{A vector of sample sizes in group 1 for two-sample mean}

\item{n2i}{A vector of sample sizes in group 2 for two-sample mean}

\item{sd1i}{A vector of standard deviations in group 1 for two-sample mean}

\item{sd2i}{A vector of standard deviations in group 2 for two-sample mean}

\item{side}{A character indicating the direction of the tested hypothesis in the primary studies (either "right" or "left")}

\item{measure}{A character indicating whether a Hedges' g should be computed based on one-sample mean or two-sample mean ("M" or "MD")}
}
\value{
Function returns a data frame with standardized effect sizes (yi), variances of these standardized effect sizes (vi), z-values (zval), and p-values as computed in primary studies (pval).
}
\description{
Function that computes Hedges' g for one-sample mean and two-sample mean and computes p-value as in primary studies.
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}

