% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GIT_print.puniformoutput_1.R
\name{print.puniformoutput}
\alias{print.puniformoutput}
\title{print.puniformoutput}
\usage{
\method{print}{puniformoutput}(x)
}
\arguments{
\item{x}{an object of class \code{puniformoutput}}
}
\value{
The print function does not return an object.
}
\description{
Print method for objecs of class \code{puniformoutput}
}
\author{
Robbie C.M. van Aert \email{R.C.M.vanAert@tilburguniversity.edu}
}

